+++ bundle/src/main/java/com/adobe/acs/commons/oak/impl/EnsureOakIndexManagerImpl.java
@@ -19,60 +19,30 @@
  */
 package com.adobe.acs.commons.oak.impl;
 
-import com.adobe.acs.commons.oak.EnsureOakIndexManager;
 import com.adobe.acs.commons.util.RequireAem;
 import com.adobe.granite.jmx.annotation.AnnotatedStandardMBean;
-import org.apache.commons.lang3.StringUtils;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.FieldOption;
 import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.osgi.service.component.annotations.ReferencePolicy;
-import org.osgi.service.component.annotations.ReferencePolicyOption;
 import org.osgi.service.metatype.annotations.AttributeDefinition;
 import org.osgi.service.metatype.annotations.Designate;
 import org.osgi.service.metatype.annotations.ObjectClassDefinition;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import javax.management.DynamicMBean;
 import javax.management.NotCompliantMBeanException;
-import javax.management.openmbean.CompositeDataSupport;
-import javax.management.openmbean.CompositeType;
 import javax.management.openmbean.OpenDataException;
-import javax.management.openmbean.OpenType;
-import javax.management.openmbean.SimpleType;
 import javax.management.openmbean.TabularData;
-import javax.management.openmbean.TabularDataSupport;
-import javax.management.openmbean.TabularType;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.CopyOnWriteArrayList;
 
 
 /**
- * This implementation of the OakIndexManager also provides a small
- * interface via the OSGI console to check the status of all
- * of EnsureOakIndex instances.
+ * This class had a circular reference, to fix this,
+ * we moved everything to EnsureOakIndexManagerExecutor and let this one only handle the ignore properties so that the PID is not lost
  */
 
-//@formatter:off
-@Component(
-        service = {DynamicMBean.class, EnsureOakIndexManager.class},
-        immediate = true,
-        property = {
-                "felix.webconsole.title=Ensure Oak Index",
-                "felix.webconsole.label=ensureOakIndex",
-                "felix.webconsole.category=Sling",
-                "jmx.objectname=com.adobe.acs.commons.oak:type=Ensure Oak Index"
-        }
-)
+@Component
 @Designate(ocd = EnsureOakIndexManagerImpl.Config.class)
-//@formatter:on
-public class EnsureOakIndexManagerImpl extends AnnotatedStandardMBean implements EnsureOakIndexManager, EnsureOakIndexManagerMBean {
+public class EnsureOakIndexManagerImpl implements EnsureOakIndexManagerProperties {
 
     //@formatter:off
     private static final String[] DEFAULT_ADDITIONAL_IGNORE_PROPERTIES = new String[]{};
@@ -99,52 +69,17 @@ public class EnsureOakIndexManagerImpl extends AnnotatedStandardMBean implements
     //@formatter:on
 
     // Disable this feature on AEM as a Cloud Service
-    @Reference(target="(distribution=classic)")
+    @Reference(target = "(distribution=classic)")
     RequireAem requireAem;
 
-    @Reference
-    EnsureOakIndexExecutor ensureOakIndexExecutor;
-
-    public EnsureOakIndexManagerImpl() throws NotCompliantMBeanException {
-        super(EnsureOakIndexManagerMBean.class);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final int ensureAll(boolean force) {
-        return ensureOakIndexExecutor.ensureAll(force);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final int ensure(final boolean force,
-                            final String ensureDefinitionPath) {
-       return ensureOakIndexExecutor.ensure(force, ensureDefinitionPath);
-    }
-
-    /**
-     * Method for displaying Ensure Oak Index state in in the MBean
-     *
-     * @return the Ensure Oak Index data in a Tabular Format for the MBean
-     * @throws OpenDataException
-     */
-    @Override
-    @SuppressWarnings("squid:S1192")
-    public final TabularData getEnsureOakIndexes() throws OpenDataException {
-        return ensureOakIndexExecutor.getEnsureOakIndexes();
-    }
 
     @Activate
     protected void activate(Config config) {
         additionalIgnoreProperties = config.properties_ignore();
     }
 
-
-    protected String[] getIgnoredProperties() {
+    @Override
+    public String[] getIgnoredProperties() {
         return Optional.ofNullable(this.additionalIgnoreProperties)
                 .map(array -> Arrays.copyOf(array, array.length))
                 .orElse(DEFAULT_ADDITIONAL_IGNORE_PROPERTIES);